created decode endpoint

+++ src/main/java/com/caesarcipher/constants/CaesarCipherConstants.java
@@ -2,7 +2,7 @@ package com.caesarcipher.constants;
 
 public class CaesarCipherConstants {
 	public static final char[] ALPHABET = "zyxwvutsrqponmlkjihgfedcba".toCharArray();
-	public static final int ALPHABET_LENGTH = ALPHABET.length - 1;
+	public static final long ALPHABET_LENGTH = ALPHABET.length - 1;
 
 	public static final String SHA1 = "SHA1";
 }
